refactored so string function are with string implementation (this will allow better unit testing).
